requirements.txt
fastapi
uvicorn
httpx
python-telegram-bot
python-dotenv
requests